refactor(core): rename file to plural
